+++ src/components/tag/tag.classnames.ts
@@ -8,9 +8,9 @@ import {
   LabelPairedCircleExclamationRegularIcon,
   LabelPairedCircleCheckRegularIcon,
   LabelPairedCircleInfoRegularIcon,
-  QuillSvgProps,
-} from '@deriv/quill-icons'
+} from '@deriv/quill-icons/LabelPaired'
 import { ExcludeNull } from 'types'
+import { QuillSvgProps } from '@deriv/quill-icons'
 
 export type BaseTagProps = ExcludeNull<
   VariantProps<typeof TagClassNamesCVA>,